refactor: rename 'subject' → 'verb object'
It had been annoying me for the longest time that I had called the
'receiver' of the verb the action 'subject' when in fact in terms of a
sentence it is the 'object'. Unfortunately 'Object' is a reserved word
so I went with 'VerbObject.

BREAKING CHANGE: if you use the TypeScript types you'll have to update
those. Otherwise there are no changes.